backend/requirements.txt
# ============================================================
# Arminda — backend dependencies
# ============================================================
# Core framework
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
psycopg[binary]==3.2.3
django-environ==0.11.2

# Multi-tenant (será ativado no Bloco 1)
django-tenants==3.7.0

# Async tasks
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.2.1

# Auth (JWT — será ativado no Bloco 1)
djangorestframework-simplejwt==5.3.1

# Auditoria
django-simple-history==3.7.0

# Filtros e paginação
django-filter==24.3

# Documentação OpenAPI
drf-spectacular==0.27.2

# Utilidades
python-dateutil==2.9.0.post0

# ============================================================
# Dev dependencies
# ============================================================
# Testes
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# Lint e formatação
ruff==0.8.4

# Debug
django-extensions==3.2.3
ipython==8.30.0
